🔨fix: 서비스단 map에서 파사드 조합 금지 규칙 반영
