+++ src/main/java/checkmo/domain/club/facade/ClubQueryFacadeImpl.java
@@ -1,5 +1,7 @@
 package checkmo.domain.club.facade;
 
+import checkmo.domain.book.facade.BookQueryFacade;
+import checkmo.domain.club.converter.ClubConverter;
 import checkmo.domain.club.entity.Club;
 import checkmo.domain.club.repository.ClubRepository;
 import checkmo.domain.club.service.query.ClubBookRecommendQueryService;
@@ -7,6 +9,7 @@ import checkmo.domain.club.service.query.ClubQueryService;
 import checkmo.domain.club.web.dto.bookshelf.BookShelfResponseDTO;
 import checkmo.domain.club.web.dto.club.ClubResponseDTO;
 import checkmo.domain.club.web.dto.meeting.MeetingResponseDTO;
+import checkmo.domain.member.facade.MemberQueryFacade;
 import checkmo.global.dto.ClubSharedDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,6 +23,9 @@ public class ClubQueryFacadeImpl implements ClubQueryFacade {
     private final ClubQueryService clubQueryService;
     private final ClubBookRecommendQueryService clubBookRecommendQueryService;
     private final ClubRepository clubRepository; // 프록시용
+
+    private final MemberQueryFacade memberQueryFacade;
+    private final BookQueryFacade bookQueryFacade;
 
     @Override
     public ClubSharedDTO.MyClubListDTO getMyClubListForShare(String memberId) {
@@ -96,7 +102,28 @@ public class ClubQueryFacadeImpl implements ClubQueryFacade {
      */
     @Override
     public ClubResponseDTO.BookRecommendListDTO getRecommendedBooks(Long clubId, Long cursorId, String memberId) {
-        return clubBookRecommendQueryService.getRecommendedBooks(clubId, cursorId, memberId);
+
+        // 1. 커서 초기화 (페이징 로직)
+        Long cursor = (cursorId == null || cursorId == 0L) ? Long.MAX_VALUE : cursorId;
+
+        // 2. ServiceImpl에서 순수 엔티티 조회
+        var bookRecommends = clubBookRecommendQueryService.getRecommendedBooks(clubId, cursor, memberId);
+
+        // 3. 외부 도메인 정보 조합 (Facade에서 처리)
+        var currentMemberNickname = memberQueryFacade.getMemberBasicInfoForShare(memberId).getNickname();
+
+        var dtoList = bookRecommends.stream()
+                .map(bookRecommend -> {
+                    var bookInfo = bookQueryFacade.getBookBasicInfoForShare(bookRecommend.getBookId());
+                    var authorInfo = memberQueryFacade.getMemberBasicInfoForShare(bookRecommend.getClubMember().getMemberId());
+                    return ClubConverter.toBookRecommendDetailDTO(bookRecommend, bookInfo, authorInfo, currentMemberNickname);
+                }).toList();
+
+        // 4. 페이징 처리 (Facade에서)
+        Long lastId = bookRecommends.isEmpty() ? null : bookRecommends.get(bookRecommends.size() - 1).getId();
+        boolean hasNext = clubBookRecommendQueryService.hasNextPage(clubId, lastId);
+
+        return ClubConverter.toBookRecommendListDTO(dtoList, hasNext, lastId);
     }
 
     /**
